fix test and type check
